Fixed size_t warnings
git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@947 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ srcphast/hst.cxx
@@ -2686,7 +2686,7 @@ void ON_ERROR_CLEANUP_AND_EXIT(void)
 void SEND_RESTART_NAME(char * name, int nchar)
 /* ---------------------------------------------------------------------- */
 {
-	int i = FileMap.size();
+	int i = (int) FileMap.size();
 	name[nchar - 1] = '\0';
 	string_trim(name);
 	std::string stdstring(name);
